Support permutations too big for transform synth
Synthesize circuits for basis translations that involve too many qubits
for their inner permutation to be represented as an array of length
`2^n`. For example, `{'0'**32, '1'**32} >> {'1'**32, '0'**32}` now
compiles. In these cases (currently, when an aligned basis translation
involves more than 10 qubits), we fall back to a simple approach that
resembles brute-forcing the truth table.

Co-authored-by: Pulkit Gupta <pgupta91@gatech.edu>